Make master branch uptoday after cvs2git migration

+++ src/com/maddyhome/idea/vim/option/NumberOption.java
@@ -1,23 +1,23 @@
 package com.maddyhome.idea.vim.option;
 
 /*
-* IdeaVim - A Vim emulator plugin for IntelliJ Idea
-* Copyright (C) 2003 Rick Maddy
-*
-* This program is free software; you can redistribute it and/or
-* modify it under the terms of the GNU General Public License
-* as published by the Free Software Foundation; either version 2
-* of the License, or (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License
-* along with this program; if not, write to the Free Software
-* Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
-*/
+ * IdeaVim - A Vim emulator plugin for IntelliJ Idea
+ * Copyright (C) 2003-2005 Rick Maddy
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ */
 
 /**
  * Represents an option with a numeric value
@@ -96,9 +96,9 @@ public class NumberOption extends TextOption
             return false;
         }
 
-        if (inRange(num.intValue()))
+        if (inRange(num))
         {
-            value = num.intValue();
+            value = num;
             fireOptionChangeEvent();
 
             return true;
@@ -123,9 +123,9 @@ public class NumberOption extends TextOption
             return false;
         }
 
-        if (inRange(value + num.intValue()))
+        if (inRange(value + num))
         {
-            value += num.intValue();
+            value += num;
             fireOptionChangeEvent();
 
             return true;
@@ -150,9 +150,9 @@ public class NumberOption extends TextOption
             return false;
         }
 
-        if (inRange(value * num.intValue()))
+        if (inRange(value * num))
         {
-            value *= num.intValue();
+            value *= num;
             fireOptionChangeEvent();
 
             return true;
@@ -177,9 +177,9 @@ public class NumberOption extends TextOption
             return false;
         }
 
-        if (inRange(value - num.intValue()))
+        if (inRange(value - num))
         {
-            value -= num.intValue();
+            value -= num;
             fireOptionChangeEvent();
 
             return true;
